+++ paper/refs.bib
@@ -28,7 +28,8 @@
   author={Dekker, Job and Alber, Frank and Aufmkolk, Sarah and Beliveau, Brian J and Bruneau, Benoit G and Belmont, Andrew S and Bintu, Lacramioara and Boettiger, Alistair and Calandrelli, Riccardo and Disteche, Christine M and others},
   journal={Molecular cell},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.molcel.2023.06.018}
 }
 
 @article{kerpedjiev2018higlass,
@@ -38,7 +39,8 @@
   volume={19},
   pages={1--12},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13059-018-1486-1}
 }
 
 @article{mair2022extricating,
@@ -49,7 +51,8 @@
   number={7911},
   pages={728--735},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-022-04718-w}
 }
 
 @article{greene2021new,
@@ -59,7 +62,8 @@
   volume={2},
   number={12},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.patter.2021.100372}
 }
 
 @software{greene2022data,
@@ -81,7 +85,8 @@ year = {2022}
   number={1},
   pages={38--51},
   year={2023},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41594-022-00892-7}
 }
 
 @article{misra2022news,
