+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper/refs.bib
 
 # Summary
 
-Jupyter Scatter is a scalable, interactive, and interlinked scatterplot widget for exploring datasets in Jupyter Notebook/Lab, Colab, and VS Code (\autoref{fig:teaser}). Thanks to its WebGL-based rendering engine [@lekschas2023regl], Jupyter Scatter can render and animate up to several million data points. The widget focuses on data-driven visual encodings and offers perceptually-effective point color and opacity settings by default. For interactive exploration, Jupyter Scatter features two-way zoom and point selections. Furthermore, the widget can compose multiple scatterplots and synchronize their views and selections, which is useful for comparing datasets. Finally, Jupyter Scatter's API integrates with Pandas DataFrames [@mckinney2010data] and Matplotlib [@hunter2007matplotlib] and offers functional methods that group properties by type to ease accessibility and readability.
+Jupyter Scatter is a scalable, interactive, and interlinked scatterplot widget for exploring datasets in Jupyter Notebook/Lab, Colab, and VS Code (\autoref{fig:teaser}). Thanks to its WebGL-based rendering engine [@lekschas2023regl], Jupyter Scatter can render and animate up to several million data points. The widget focuses on data-driven visual encodings and offers perceptually-effective point color and opacity settings by default. For interactive exploration, Jupyter Scatter features two-way zoom and point selections. Furthermore, the widget can compose multiple scatterplots and synchronize their views and selections, which is useful for comparing datasets. Finally, Jupyter Scatter's API integrates with Pandas DataFrames [@mckinney2010data] and Matplotlib [@hunter2007matplotlib] and offers functional methods that group properties by type to ease accessibility and readability. Extensive documentation and how tos can be found at [https://jupyter-scatter.dev](https://jupyter-scatter.dev) and the code is available at [https://github.com/flekschas/jupyter-scatter](https://github.com/flekschas/jupyter-scatter).
 
 ![Examples of Jupyter Scatter. Top row left to right: A 10M point scatterplot of the Roessler Attractor. A connected scatterplot of the market capitalization over the last five years of top ten S&P500 company according to YCharts. Five linked embedding plots of epigenomic data [@dekker2023spatial] that are connected to the HiGlass genome browser [@kerpedjiev2018higlass]. Bottom row left to right: A single-cell embedding plot of tumor data [@mair2022extricating] that was clustered and annotated with FAUST [@greene2021new, @greene2022data]. Several linked embedding plots of chromatin state datasets [@spracklin2023diverse]. An embedding plot of news headlines [@misra2022news] that is linked to a widget for displaying selected articles.
 \label{fig:teaser}](paper/teaser.jpg)
