### 2024-06-11 3858096
fix: preserve filter state upon changing visual encoding
The main fix was in https://github.com/flekschas/regl-scatterplot/pull/181
