### 2024-06-14 #6867 @crvernon
Just adding another ping.

> @editorialbot invite @faroit as editor
>
> 👋 @faroit - can you take this one on as editor?
