### 2024-06-14 f12a4b6
refactor: improve view reset behavior
