### 2024-06-20 6f3ff73
refactor: adjust `scatter.data()` to make view reset and animation explicit
