+++ jscatter/jscatter.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 
 from matplotlib.colors import to_rgba, Normalize, LogNorm, PowerNorm, LinearSegmentedColormap, ListedColormap
-from typing import Dict, Optional, Union, List, Tuple
+from typing import Dict, Optional, Union, List, Tuple, Literal
 
 from .encodings import Encodings
 from .widget import JupyterScatter, SELECTION_DTYPE
@@ -452,6 +452,8 @@ class Scatter():
         self,
         data: Optional[pd.DataFrame] = None,
         use_index: Optional[Union[bool, Undefined]] = UNDEF,
+        reset_view: Optional[Union[bool, Undefined]] = UNDEF,
+        animate: Optional[Union[bool, Undefined]] = UNDEF,
         **kwargs
     ) -> Union[Scatter, dict]:
         """
@@ -460,13 +462,19 @@ class Scatter():
         Parameters
         ----------
         data : pd.DataFrame, optional
-            The data frame that holds the x and y coordinates as well as other
-            possible dimensions that can be used for color, size, or opacity
-            encoding.
+            The new data frame that holds the x and y coordinates as well as
+            other possible dimensions that can be used for color, size, or
+            opacity encoding.
         label_index : bool, optional
             If `True` and if the data frame's index is not an instance of
             `RangeIndex` then the selection and filter methods will reference
             points by the data frame's label index instead of the row index.
+        view_data : bool, optional
+            If `True`, the view will adjust to the new data extent. If `False`,
+            the view will remain unchanged.
+        animate_change : bool, optional
+            If `True` and if the number of points (i.e., rows in the data frame)
+            are the same, the points will transition smoothly.
 
         Returns
         -------
@@ -493,6 +501,8 @@ class Scatter():
         {'data': <pandas.DataFrame>, 'use_label_index': False}
         """
         if data is not None:
+            old_n = self._n
+
             try:
                 self._n = len(data)
                 self._data = data
@@ -504,23 +514,32 @@ class Scatter():
 
             self._points = np.zeros((self._n, 6))
 
-            # Reset scale & norm ranges
-            self._x_scale.vmin = None
-            self._x_scale.vmax = None
-            self._y_scale.vmin = None
-            self._y_scale.vmax = None
-            self._color_norm.vmin = None
-            self._color_norm.vmax = None
-            self._opacity_norm.vmin = None
-            self._opacity_norm.vmax = None
-            self._size_norm.vmin = None
-            self._size_norm.vmax = None
-            self._connection_color_norm.vmin = None
-            self._connection_color_norm.vmax = None
-            self._connection_opacity_norm.vmin = None
-            self._connection_opacity_norm.vmax = None
-            self._connection_size_norm.vmin = None
-            self._connection_size_norm.vmax = None
+            if reset_view == True:
+                self._x_scale.vmin = None
+                self._x_scale.vmax = None
+                self._y_scale.vmin = None
+                self._y_scale.vmax = None
+                self._color_norm.vmin = None
+                self._color_norm.vmax = None
+                self._opacity_norm.vmin = None
+                self._opacity_norm.vmax = None
+                self._size_norm.vmin = None
+                self._size_norm.vmax = None
+                self._connection_color_norm.vmin = None
+                self._connection_color_norm.vmax = None
+                self._connection_opacity_norm.vmin = None
+                self._connection_opacity_norm.vmax = None
+                self._connection_size_norm.vmin = None
+                self._connection_size_norm.vmax = None
+
+            if 'skip_widget_update' not in kwargs:
+                # We have to update the following widget props now to ensure
+                # that the update propagated to the widget prior to updating the
+                # points below. This seems to be a bug in Jupyter Lab but I
+                # noticed that the order in which updates are registered on the
[…]

[…]
