+++ paper/refs.bib
@@ -214,6 +214,26 @@ year = {2022}
   url = {https://www.geonames.org},
 }
 
+@article{vanderplas2018altair,
+  doi = {10.21105/joss.01057},
+  url = {https://doi.org/10.21105/joss.01057},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {32},
+  pages = {1057},
+  author = {Jacob VanderPlas and Brian Granger and Jeffrey Heer and Dominik Moritz and Kanit Wongsuphasawat and Arvind Satyanarayan and Eitan Lees and Ilia Timofeev and Ben Welsh and Scott Sievert},
+  title = {Altair: Interactive Statistical Visualizations for Python},
+  journal = {Journal of Open Source Software}
+}
+
+@misc{pyobsplot,
+  author = {Julien Barnier},
+  title = {{Observable Plot} in {Jupyter} notebooks and {Quarto} documents},
+  year = {2024},
+  url = {https://github.com/juba/pyobsplot},
+}
+
 @article{manz2024general,
   title = {A General Framework for Comparing Embedding Visualizations Across Class-Label Hierarchies},
   url = {https://osf.io/puxnf},
