### 2024-06-21 9c5c86f
Add URLs and whitespaces to all papers. Fix Okabe Ito reference
