+++ paper.md
@@ -55,7 +55,7 @@ scatter.show()
 
 ![GeoNames Dataset of Cities Around the World.\label{fig:geonames}](paper/geonames.jpg)
 
-Without specifying a color map, Jupyter Scatter uses the categorical colorblind-safe palette from @okabe2008color for the `Continent` column, which has seven unique values. For columns with continuous data, it automatically selects Matplotlib's [@hunter2007matplotlib] _Viridis_ color palette. As shown in \autoref{fig:teaser} and \autoref{fig:geonames} Left, Jupyter Scatter dynamically adjusts the point opacity based on the point density within the field of view. This means points become more opaque when zooming into sparse areas and more transparent when zooming out into an area that contains many points. The dynamic opacity addresses over-plotting issues when zoomed out and visibility issues when zoomed in.
+Without specifying a color map, Jupyter Scatter uses the categorical colorblind-safe palette from @okabe2002color for the `Continent` column, which has seven unique values. For columns with continuous data, it automatically selects Matplotlib's [@hunter2007matplotlib] _Viridis_ color palette. As shown in \autoref{fig:teaser} and \autoref{fig:geonames} Left, Jupyter Scatter dynamically adjusts the point opacity based on the point density within the field of view. This means points become more opaque when zooming into sparse areas and more transparent when zooming out into an area that contains many points. The dynamic opacity addresses over-plotting issues when zoomed out and visibility issues when zoomed in.
 
 Jupyter Scatter offers many ways to customize the point color, size, and opacity encodings. To simplify configuration, it provides topic-specific methods for setting up the scatterplot, rather than requiring all properties to be set during the instantiation of `Scatter`. For instance, as shown in \autoref{fig:geonames} Right, the point opacity (`0.5`), size (asinh-normalized), and color (log-normalized population using Matplotlib's [@hunter2007matplotlib] _Magma_ color palette in reverse order) can be set using the following methods.
 
