+++ paper/refs.bib
@@ -20,57 +20,61 @@
   pages = {90--95},
   publisher = {IEEE COMPUTER SOC},
   doi = {10.1109/MCSE.2007.55},
-  year = 2007
+  year = {2007},
+  url = {https://doi.org/10.1109/MCSE.2007.55},
 }
 
 @article{dekker2023spatial,
-  title={Spatial and temporal organization of the genome: Current state and future aims of the 4D nucleome project},
-  author={Dekker, Job and Alber, Frank and Aufmkolk, Sarah and Beliveau, Brian J and Bruneau, Benoit G and Belmont, Andrew S and Bintu, Lacramioara and Boettiger, Alistair and Calandrelli, Riccardo and Disteche, Christine M and others},
-  journal={Molecular cell},
-  year={2023},
-  publisher={Elsevier},
-  doi={10.1016/j.molcel.2023.06.018}
+  title = {Spatial and temporal organization of the genome: Current state and future aims of the 4D nucleome project},
+  author = {Dekker, Job and Alber, Frank and Aufmkolk, Sarah and Beliveau, Brian J and Bruneau, Benoit G and Belmont, Andrew S and Bintu, Lacramioara and Boettiger, Alistair and Calandrelli, Riccardo and Disteche, Christine M and others},
+  journal = {Molecular cell},
+  year = {2023},
+  publisher = {Elsevier},
+  doi = {10.1016/j.molcel.2023.06.018},
+  url = {https://doi.org/10.1016/j.molcel.2023.06.018},
 }
 
 @article{kerpedjiev2018higlass,
-  title={HiGlass: web-based visual exploration and analysis of genome interaction maps},
-  author={Kerpedjiev, Peter and Abdennur, Nezar and Lekschas, Fritz and McCallum, Chuck and Dinkla, Kasper and Strobelt, Hendrik and Luber, Jacob M and Ouellette, Scott B and Azhir, Alaleh and Kumar, Nikhil and others},
-  journal={Genome biology},
-  volume={19},
-  pages={1--12},
-  year={2018},
-  publisher={Springer},
-  doi={10.1186/s13059-018-1486-1}
+  title = {HiGlass: web-based visual exploration and analysis of genome interaction maps},
+  author = {Kerpedjiev, Peter and Abdennur, Nezar and Lekschas, Fritz and McCallum, Chuck and Dinkla, Kasper and Strobelt, Hendrik and Luber, Jacob M and Ouellette, Scott B and Azhir, Alaleh and Kumar, Nikhil and others},
+  journal = {Genome biology},
+  volume = {19},
+  pages = {1--12},
+  year = {2018},
+  publisher = {Springer},
+  doi = {10.1186/s13059-018-1486-1},
+  url = {https://doi.org/10.1186/s13059-018-1486-1},
 }
 
 @article{mair2022extricating,
-  title={Extricating human tumour immune alterations from tissue inflammation},
-  author={Mair, Florian and Erickson, Jami R and Frutoso, Marie and Konecny, Andrew J and Greene, Evan and Voillet, Valentin and Maurice, Nicholas J and Rongvaux, Anthony and Dixon, Douglas and Barber, Brittany and others},
-  journal={Nature},
-  volume={605},
-  number={7911},
-  pages={728--735},
-  year={2022},
-  publisher={Nature Publishing Group UK London},
-  doi={10.1038/s41586-022-04718-w}
+  title = {Extricating human tumour immune alterations from tissue inflammation},
[…]
