### 2024-06-21 41e2972
chore: Remove Python 3.7 and add 3.12. Bump default dev version to 3.12
