+++ pyproject.toml
@@ -15,11 +15,11 @@ classifiers = [
     "Intended Audience :: Developers",
     "Intended Audience :: Science/Research",
     "Topic :: Multimedia :: Graphics",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 keywords = [
     "scatter",
@@ -81,7 +81,7 @@ path = "js"
 
 [tool.hatch.envs.default]
 features = ["dev"]
-python = "3.11"
+python = "3.12"
 
 [tool.hatch.envs.default.scripts]
 test = "pytest ."
