### 2024-06-21 b4423a4
refactor: finalize `data()` signature
