### 2024-06-21 e1c3c3c
fix: adjust test to new `data()` signature
