### 2024-06-23 a7619e4
Update CITATION.cff
