### 2024-06-23 ef1dc88
Update CITATION.cff
