+++ CITATION.cff
@@ -22,7 +22,8 @@ preferred-citation:
   publisher: "arXiv"
   journal: "arXiv"
   year: 2024
-  month: 5
+  month: 6
+  day: 20
   title: "Jupyter Scatter: Interactive Exploration of Large-Scale Datasets"
   url: "https://arxiv.org/abs/2406.14397"
 title: "Jupyter Scatter: Interactive Exploration of Large-Scale Datasets"
