### 2024-06-23 e839c43
chore: bump anywidget min version to `0.9.0`
