+++ js/src/index.js
@@ -2397,7 +2397,7 @@ function modelWithSerializers(model, serializers) {
   }
 }
 
-export async function render({ model, el }) {
+async function render({ model, el }) {
   const view = new JupyterScatterView({
     el: el,
     model: modelWithSerializers(model, {
@@ -2411,3 +2411,5 @@ export async function render({ model, el }) {
   view.render();
   return () => view.destroy();
 }
+
+export default { render };
