+++ pyproject.toml
@@ -32,7 +32,7 @@ keywords = [
     "widgets",
 ]
 dependencies = [
-    "anywidget>=0.2.0",
+    "anywidget>=0.9.0",
     "ipywidgets>=7.6,<9",
     "ipython",
     "matplotlib",
