### 2024-06-23 81a27ea
fix: reset scale & norm ranges in `scatter.data()` (#132)
* fix: reset scale & norm ranges in `scatter.data()`

* fix: update x/y (scale) domains upon data updating

* test: check the widget rather than the scatter state
