### 2024-06-23 8c0c532
fix: reset scale & norm ranges in `scatter.data()` (#132)
* fix: reset scale & norm ranges in `scatter.data()`

* fix: update x/y (scale) domains upon data updating

* test: check the widget rather than the scatter state
