### 2024-06-23 f51090a
fix: `scatter.axes(labels=['x_label', 'y_label'])`
