### 2024-06-23 9b096a9
fix: `scatter.axes(labels=['x_label', 'y_label'])` (#140)
* fix: `scatter.axes(labels=['x_label', 'y_label'])`

* docs: update changelog
