### 2024-06-25 ff75acd
feat: add `scatter.show_tooltip()` (#139)
* feat: add `scatter.show_tooltip(point_idx)`

This is a little helper method to programmatically show a tooltip.

* docs: update changelog
