+++ js/src/index.js
@@ -385,8 +385,9 @@ class JupyterScatterView {
 
   customEventHandler(event) {
     if (event.type === this.eventTypes.TOOLTIP) {
-      if (event.index !== this.tooltipPointIdx) return;
+      if (event.index !== this.tooltipPointIdx && event.show !== true) return;
       this.tooltipDataHandlers(event)
+      if (event.show) this.showTooltip(event.index);
     }
     if (event.type === this.eventTypes.VIEW_RESET) {
       if (!this.scatterplot) return;
