+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## v0.16.2
+## v0.17.0
 
+- Feat: Add `scatter.show_tooltip(point_idx)`
 - Fix: reset scale & norm ranges upon updating the data via `scatter.data()`
 - Fix: ensure `scatter.axes(labels=['x_label', 'y_label'])` works properly ([#137](https://github.com/flekschas/jupyter-scatter/issues/137]))
 
