### 2024-07-01 347461a
fix: typo prevents tooltip from showing up
Fix #141
