### 2024-07-05 #6867 @faroit
👋 Hi @flekschas, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

Regarding this submission, it would be great to have reviewers with domain knowledge who also have experience with visualization software. In addition to your proposals I would like to ask the following potential reviewers

- @`j9ac9k`
- @`campagnola`
- @`csadorf`
- @`ianhi`

For this, please check the JOSS [COI policy](<https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy>) and get back to me if i could invite them.
