### 2024-07-19 #6867 @faroit
> Thanks @faroit. I don't see any conflict of interest for the reviewers you suggested. However, one small note: I acted as a reviewer for `ianhi`'s recent JOSS paper `mpl-interactions`: https://joss.theoj.org/papers/10.21105/joss.05651. I don't think this counts as a conflict of interest but I want to mention it anyway.

thanks for mentioning that, I would rather tend to treat this as a potential COI, so I would request without that reviewer for now.
