### 2024-07-19 #6867 @xiaohk
> 👋 @j9ac9k @campagnola @csadorf @xiaohk @Fil @paulbrodersen - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is https://github.com/flekschas/jupyter-scatter
>
> This issue is a "pre-review" issue in which reviewers are assigned. Once sufficient reviewers are recruited we will open a dedicated review issue where the review will take place.

Hi @faroit! Happy to review this submission!
