### 2024-07-30 #7059 @editorialbot
[REVIEW]: Jupyter Scatter: Interactive Exploration of Large-Scale Datasets
**Submitting author:** @flekschas (<a href="http://orcid.org/0000-0001-8432-4835">Fritz Lekschas</a>)
**Repository:** https://github.com/flekschas/jupyter-scatter/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.18.0
