### 2024-08-01 #7059 @paulbrodersen
Jupyter Scatter is an excellent tool for visualizing embeddings of large scale data. It comes with great documentation and mostly works out-of-the-box. Barring very minor revisions, I would hence readily support acceptance of this submission.
While most review items are well covered, the area I would be most keen on seeing improvements in is the description of the state of the field, in particular, a discussion of software alternatives, and how Jupyter Scatter compares to them.

Below, I have split my detailed comments into two sections.

1. "Checklist items" explains why I left some of the boxes on the reviewer checklist unchecked. I have mostly adhered to the order in the checklist, but merged my comments to "Installation" and "Installation instructions" into one. As the work presented here is very good, most comments are to be understood as gentle nudges. Essential issues are marked in bold (but those are all easily addressed).
2. "Additional comments" are just that: comments that occurred to me while reviewing the work, not a TODO list.

# Checklist items

## Installation & Installation instructions

  - jupyterlab is only an optional dependency of Jupyter Scatter. While I understand that Jupyter Scatter can be made to work with other notebooks, I would suggest adding it to the list of "proper" dependencies (and pin version >=2) in line with python's "batteries included" philosophy. Installation of non-standard setups could be documented separately, as it is already being done for jupyterlab<2 in the README, and for google colab in a notebook. You may want to standardize how you document different installation instructions / non-standard setups (or drop support for them entirely).
  - **Installation from source fails if nodejs & npm are not installed.** I would suggest adding separate instructions for installing from source in the README, and noting the (currently implicit) nodejs/npm requirement.

## Example usage

The authors provide excellent example notebooks. However, I ran into two issues when trying to run them:

- examples.ipynb:
  Visualizing the Roessler attractor with up to 100 000 points results in very pretty figures. Using 1 000 000 points (the value used in the notebook) just results in a blank screen on my machine (and yes, I waited until the cell evaluation was complete). I would suggest reducing the default value to 10-100k points to ensure that the figure is drawn properly.
- **linking.ipynb: `pyarrow` is not a listed dependency of Jupyter Scatter but required to evaluate cell #4.** At the very least document the dependency in the notebook. Alternatively, add the dependency to the pyproject.toml, potentially in its own group, e.g. [examples].

## Functionality documentation

While the API documentation is good, it is easily missed as the link to it is only in the README index, not the README body. From a UI perspective, I would expect an index to only direct me to parts of the same document (i.e. other sections of the README), not to a separate site. I missed the API documentation on first and second read, and would hence strongly suggest to have an explicit section in the README along the lines of:

``` markdown
## Docs

The online documentation including examples and a complete API description can be found [here](https://jupyter-scatter.dev/).
```

Ditto for the "Examples" item in the index. [Explicit is better than implicit.](https://peps.python.org/pep-0020/)

## Automated tests

  - There is no documentation on how to run the pytest test suite (though the project follows standard best practices, so anyone interested in testing it, probably can). **To fully satisfy this journal's requirements, it wouldn't hurt to add explicit testing instructions, e.g. in the CONTRIBUTING.md.**
  - The coverage of jscatter.py seems low, but one would need to do more digging than I am willing to do to find out if coverage lacks in areas that matter. Automated testing of visual outputs is hard and often brittle, so I am willing to give the authors the benefit of the doubt that they are testing what can and/or should be tested.

  ```
  ---------- coverage: platform linux, python 3.12.4-final-0 -----------
  Name                     Stmts   Miss  Cover
  --------------------------------------------
  jscatter/__init__.py        11      4    64%
  jscatter/color_maps.py       4      0   100%
  jscatter/compose.py        138    128     7%
  jscatter/encodings.py      104      3    97%
  jscatter/jscatter.py      1564    568    64%
  jscatter/types.py           49      0   100%
  jscatter/utils.py          102     20    80%
  jscatter/widget.py         219     63    71%
  --------------------------------------------
  TOTAL                     2191    786    64%
```

## Summary

The summary is good if a bit jargon-heavy. Specifically, the word "widget" is used extensively without ever being explained or defined. In my personal experience, the word "widget" is not widely understood outside of data visualisation circles.

## State of the field

The authors describe well how Jupyter Scatter relates to the popular general-purpose python plotting libraries. Arguably, these are not the viable alternatives for large scale & interactive visualizations, generating the need for specialised tools like Jupyter Scatter. **However, the paper would benefit from a candid discussion of more direct competitors.** For example, section 2.2 in [this paper](https://arxiv.org/abs/2306.09328) could serve as a starting point for a more in-depth discussion of the pros and cons of alternative software solutions (they even cite regl-scatterplot, which is -- as I understand it -- the basis of the current project!).

> Finally, except for Plotly, none of the tools support interactive point selections, a key feature of Jupyter Scatter.

The meaning of the word "support" here requires qualification. For example, Matplotlib (and hence Seaborn) support pick events specifically, and the documentation of holoviews and bokeh suggest that they do, too, so arguably, these libraries "support" interactive point selection as well.

# Additional comments

## License

@flekschas is listed as the sole copyright holder. Is that correct?

## Source code

This a review of a journal submission, and not a code review. However, if I were to put on my code review hat for just one second and make only a single suggestion, then that suggestion would be to break up the main Scatter class: it is simply too large, and the resulting cognitive load may deter potential contributors. Otherwise the code is very readable with expressive variable names, extensive type hinting, and good formatting (black, presumably?).

## Branding

It is a bit late for the purpose of this submission, but I can't help but wonder if it were better to re-brand the project as a widget for visualizing large scale embeddings, rather than a general multi-purpose scatter plot generator.
All the finer implementation details work towards that goal (plot synchronization, tooltips, and even the data density driven opacity selection), and tellingly, all non-trivial examples deal with visualizing embedding data as well.
