### 2024-08-02 a6db59d
docs: make link to website more prominent
Also make sure the index only links to parts of the README.
