+++ README.md
@@ -43,13 +43,25 @@ Imagine trying to explore a dataset of millions of data points as a 2D scatter. 
 
 Internally, Jupyter Scatter uses [regl-scatterplot](https://github.com/flekschas/regl-scatterplot/) for WebGL rendering, [traitlets](https://traitlets.readthedocs.io/en/stable/) for two-way communication between the JS and iPython kernels, and [anywidget](https://anywidget.dev/) for composing the widget.
 
+**Docs**
+
+Visit [https://jupyter-scatter.dev](https://jupyter-scatter.dev) for detailed documentation including examples and a complete API description.
+
+---
+
 **Index**
 
 1. [Install](#install)
 2. [Get Started](#get-started)
-3. [Docs](https://jupyter-scatter.dev)
-4. [Examples](notebooks/examples.ipynb)
-5. [Development](#development)
+   1. [Simplest Example](#simplest-example)
+   2. [Pandas DataFrame Example](#pandas-dataframe-example)
+   3. [Advanced Example](#advanced-example)
+   4. [Functional API Example](#functional-api-example)
+   5. [Linking Scatter Plots](#linking-scatter-plots)
+   6. [Visualize Millions of Data Points](#visualize-millions-of-data-points)
+   7. [Google Colab](#google-colab)
+4. [Development](#development)
+5. [Citation](#citation)
 
 ## Install
 
@@ -70,7 +82,7 @@ For a minimal working example, take a look at [test-environments](test-environme
 > [!TIP]
 > Visit [jupyter-scatter.dev](https://jupyter-scatter.dev) for details on all essential features of Jupyter Scatter and check out our [full-blown tutorial](https://github.com/flekschas/jupyter-scatter-tutorial) from SciPy '23.
 
-#### Simplest Example
+### Simplest Example
 
 In the simplest case, you can pass the x/y coordinates to the plot function as follows:
 
@@ -86,7 +98,7 @@ jscatter.plot(x, y)
 
 <img width="448" alt="Simplest scatter plotexample" src="https://user-images.githubusercontent.com/932103/116143120-bc5f2280-a6a8-11eb-8614-51def74d692e.png">
 
-#### Pandas Example
+### Pandas DataFrame Example
 
 Say your data is stored in a Pandas dataframe like the following:
 
@@ -119,7 +131,7 @@ jscatter.plot(data=df, x='mass', y='speed')
 <img width="448" alt="Pandas scatter plot example" src="https://user-images.githubusercontent.com/932103/116143383-1364f780-a6a9-11eb-974c-4facec249974.png">
 </details>
 
-#### Advanced example
+### Advanced Example
 
 Often you want to customize the visual encoding, such as the point color, size, and opacity.
 
@@ -144,7 +156,7 @@ Also notice how jscatter uses an appropriate color map by default based on the d
 
 You can, of course, customize the color map and many other parameters of the visual encoding as shown next.
 
-#### Functional API Example
+### Functional API Example
 
 The [flat API](#advanced-example) can get overwhelming when you want to customize a lot of properties. Therefore, jscatter provides a functional API that groups properties by type and exposes them via meaningfully-named methods.
 
@@ -167,7 +179,7 @@ Moreover, all arguments are optional. If you specify arguments, the methods will
 
 Finally, the scatter plot is interactive and supports two-way communication. Hence, if you select some point with the lasso tool and then call `scatter.selection()` you will get the current selection.
 
-#### Linking Scatter Plots
+### Linking Scatter Plots
 
 To explore multiple scatter plots and have their view, selection, and hover interactions link, use `jscatter.link()`.
 
[…]
