### 2024-08-03 7fb684d
docs: add note regarding Node dependency when building from source
