### 2024-08-04 07d0257
ci: remove python setup from JS build and have npm verify cache
Hopefully this will fix the odd `npm warn tarball tarball data for ... seems to be corrupted. Trying again.` error
