### 2024-08-04 b6f1880
ci: test updating `npm` to fix the install error
https://stackoverflow.com/a/52853169
