+++ .github/workflows/build.yml
@@ -51,6 +51,6 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
       - run: |
-          npm cache verify
+          npm update
           npm ci
         working-directory: js
