+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
           node-version: "20.x"
       - run: |
           cd js
-          npm cache verify
+          npm update
           npm ci
 
       - uses: actions/setup-python@v5
