### 2024-08-04 12ceebb
fix: bump regl-scatterplot to v1.10.2 (#143)
* fix: bump regl-scatterplot to v1.10.2

There was an issue with rendering more than 1M points with regl-scatterplot due to a bug introduced in v1.9.0.

* ci: remove python setup from JS build and have npm verify cache

Hopefully this will fix the odd `npm warn tarball tarball data for ... seems to be corrupted. Trying again.` error

* ci: test updating `npm` to fix the install error

https://stackoverflow.com/a/52853169
