### 2024-08-04 31f5d9d
docs: add Node requirement to development section.
Also add a note regarding anywidget's HMR
