### 2024-08-04 b97e4c2
docs: add a note about how to run tests
