### 2024-08-04 42c97ff
test: fix pandas deprecation warning and remove unused `tooltip_histograms_bins` argument
