### 2024-08-05 8170acf
Remove extensive use of `widget` in Summary and clarify "support for interactive point selections"
