### 2024-08-06 e229f1a
Update anywidget citation to point to the preprint
