+++ paper/refs.bib
@@ -208,11 +208,14 @@ year = {2022}
   url = {https://github.com/ipython/traitlets},
 }
 
-@misc{anywidget,
-  author = {Manz, Trevor},
-  title = {{Anywidget}: {Jupyter} Widgets Made Easy},
-  year = {2024},
-  url = {https://github.com/manzt/anywidget},
+@article{anywidget,
+ title = {anywidget: reusable widgets for interactive analysis and visualization in computational notebooks},
+ url = {osf.io/tw9sg},
+ doi = {10.31219/osf.io/tw9sg},
+ publisher = {OSF Preprints},
+ author = {Manz, Trevor and Abdennur, Nezar and Gehlenborg, Nils},
+ year = {2024},
+ month = {Jun}
 }
 
 @misc{geonames,
