### 2024-08-06 #7059 @flekschas
Thank you for the detailed and thorough review! We appreciate your insights and have initiated PRs to address the issues raised. Below, I will respond to each point:

**Installation & Installation instructions**
1. _`jupyterlab` dependency_: While we understand and appreciate the notion of "batteries included," `jupyterlab` is not a strict dependency. Jupyter Scatter, thanks to `anywidget`, runs in various Notebook-like environments (and even [non-notebook environments like dashboards](https://github.com/flekschas/jupyter-scatter/issues/124)). Jupyter Lab is one of them. While it's a popular environment and one we prefer for development, adding Jupyter Lab as a proper dependency would mean that in all other environments (Jupyter Notebook, Google Colab, VSCode, etc.), users would install an unnecessary package.
2. _Source installation_: We have added a note to the _Install_ and _Development_ sections in the `README` to highlight the need for `Node`. [PR #142](https://github.com/flekschas/jupyter-scatter/pull/142)

**Example usage**
1. `examples.ipynb`: Thank you for bringing this issue to our attention. It turned out to be a bug in a recent version of `regl-scatterplot`. We have updated `regl-scatterplot` and released a new patch version of Jupyter Scatter (v0.17.2), which fixes the issue. You can now render millions of points as expected again. [PR #143](https://github.com/flekschas/jupyter-scatter/pull/143)
2. `linking.ipynb`: We have updated the notebook to include a `pip install pyarrow` call. [PR #144](https://github.com/flekschas/jupyter-scatter/pull/144)

**Functionality documentation**
1. _Link to docs_: We have added an extra sentence prior to the _Index_ in the `README` to make the documentation more findable. [PR #142](https://github.com/flekschas/jupyter-scatter/pull/142)

**Automated tests**
1. _Running tests_: We have added a note to the _Development_ section in the `README` on how to run tests. This seems the most fitting location as anyone wanting to contribute to Jupyter Scatter will likely read through the development setup.
2. _Test coverage_: We agree that the test coverage could be higher and plan to write more tests in upcoming releases.

**Summary**
1. _Jargon_: We agree we never really explained what a widget is. To reduce jargon, we have removed the term from the summary. [PR #145](https://github.com/flekschas/jupyter-scatter/pull/145)

**State of the field**
1. _Discussion of direct competitors_: Jupyter Scatter is a Python-based general-purpose scatterplot widget for Notebook-like environments, optimized for interactive exploration of large-scale datasets. The referenced WizMap paper discusses embedding visualizations outside the context of Notebook-like environments. These are not direct competitors because Jupyter Scatter is a general-purpose tool for rendering scatterplots in Notebook-like environments. None of the tools mentioned in the WizMap paper section 2.2 run in a notebook, except WizMap itself, which is only meant for visualizing embeddings. Therefore, we have compared Jupyter Scatter against other Python-based general-purpose visualization tools that run in Notebook-like environments. If you have specific tools in mind that we may have missed, please let us know.
2. _Support interactive point selections_: Thank you for pointing this out. We have clarified that we are referring to interactive **two-way** point selections. It's one thing to be able to select points visually in the plot, but without being able to use that selection for follow-up analysis, the point selection has limited benefit for data analysis. We have clarified the language to make this point more clearly. [PR #145](https://github.com/flekschas/jupyter-scatter/pull/145)

**Additional comments**
1. _License_: Yes, that's correct.
2. _Source code_: We agree! The `Scatter` class has gotten too large and should ideally be broken up. It is on our to-do list, but it will take some time to properly refactor everything.
3. _Branding_: We respectfully disagree. While many examples involve embeddings because Trevor and I work extensively with embeddings, there's nothing in Jupyter Scatter that is specifically tailored to embeddings. The implementation details aim for interactive, scalable, and comparable scatter plots, useful in various contexts. For example, tooltips are beneficial for showing additional dimensions of a data point, plot synchronization aids in comparison, and density-driven point opacity enhances large-scale scatter plot interactivity, avoiding over-plotting and perception issues.

Thank you again for your valuable feedback! We hope this addresses the issues you raised.
