+++ jscatter/jscatter.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 
 from matplotlib.colors import to_rgba, Normalize, LogNorm, PowerNorm, LinearSegmentedColormap, ListedColormap
-from typing import Dict, Optional, Union, List, Tuple
+from typing import Dict, Optional, Union, List, Tuple, Literal
 
 from .encodings import Encodings
 from .widget import JupyterScatter, SELECTION_DTYPE
@@ -452,6 +452,9 @@ class Scatter():
         self,
         data: Optional[pd.DataFrame] = None,
         use_index: Optional[Union[bool, Undefined]] = UNDEF,
+        reset_scales: Optional[bool] = False,
+        reset_view: Optional[bool] = False,
+        animate: Optional[bool] = False,
         **kwargs
     ) -> Union[Scatter, dict]:
         """
@@ -460,13 +463,22 @@ class Scatter():
         Parameters
         ----------
         data : pd.DataFrame, optional
-            The data frame that holds the x and y coordinates as well as other
-            possible dimensions that can be used for color, size, or opacity
-            encoding.
+            The new data frame that holds the x and y coordinates as well as
+            other possible dimensions that can be used for color, size, or
+            opacity encoding.
         label_index : bool, optional
             If `True` and if the data frame's index is not an instance of
             `RangeIndex` then the selection and filter methods will reference
             points by the data frame's label index instead of the row index.
+        reset_scales : bool, optional
+            If `True`, all scales (and norms) will be reset to the extend of the
+            the new data.
+        reset_view : bool, optional
+            If `True`, the view will be reset to the origin.
+        animate : bool, optional
+            If `True`, if the number of points remain the same, and if
+            `reset_scales` is `False`, the points will transition smoothly. If
+            `reset_view` is `True`, the view will also transition smoothly.
 
         Returns
         -------
@@ -493,6 +505,8 @@ class Scatter():
         {'data': <pandas.DataFrame>, 'use_label_index': False}
         """
         if data is not None:
+            old_n = self._n
+
             try:
                 self._n = len(data)
                 self._data = data
@@ -503,6 +517,35 @@ class Scatter():
                 return self
 
             self._points = np.zeros((self._n, 6))
+
+            same_n = old_n == self._n
+
+            if reset_scales == True:
+                self._x_scale.vmin = None
+                self._x_scale.vmax = None
+                self._y_scale.vmin = None
+                self._y_scale.vmax = None
+                self._color_norm.vmin = None
+                self._color_norm.vmax = None
+                self._opacity_norm.vmin = None
+                self._opacity_norm.vmax = None
+                self._size_norm.vmin = None
+                self._size_norm.vmax = None
+                self._connection_color_norm.vmin = None
+                self._connection_color_norm.vmax = None
+                self._connection_opacity_norm.vmin = None
+                self._connection_opacity_norm.vmax = None
+                self._connection_size_norm.vmin = None
+                self._connection_size_norm.vmax = None
+
+            if 'skip_widget_update' not in kwargs:
+                # We have to update the following widget props now to ensure
+                # that the update propagated to the widget prior to updating the
+                # points below. This seems to be a bug in Jupyter Lab but I
+                # noticed that the order in which updates are registered on the
+                # front-end does *not* correspond to the call order in Python!
+                self.update_widget('transition_points', animate and same_n)
+                self.update_widget('prevent_filter_reset', False)
 
             self.x(skip_widget_update=True, **self.x())
             self.y(skip_widget_update=True, **self.y())
@@ -515,12 +558,25 @@ class Scatter():
             self.connection_size(skip_widget_update=True, **self.connection_size())
 
             if 'skip_widget_update' not in kwargs:
[…]
