+++ jscatter/widget.py
@@ -13,7 +13,10 @@ from traittypes import Array # type: ignore
 from .utils import to_hex, with_left_label
 
 SELECTION_DTYPE = 'uint32'
-TOOLTIP_EVENT_TYPE = 'tooltip';
+EVENT_TYPES = {
+    "TOOLTIP": "tooltip",
+    "VIEW_RESET": "view_reset",
+}
 
 def component_idx_to_name(idx):
     if idx == 2:
@@ -57,6 +60,7 @@ class JupyterScatter(anywidget.AnyWidget):
 
     # Data
     points = Array(default_value=None).tag(sync=True, **ndarray_serialization)
+    transition_points = Bool(False).tag(sync=True)
     prevent_filter_reset = Bool(False).tag(sync=True)
     selection = Array(default_value=None, allow_none=True).tag(sync=True, **ndarray_serialization)
     filter = Array(default_value=None, allow_none=True).tag(sync=True, **ndarray_serialization)
@@ -213,7 +217,6 @@ class JupyterScatter(anywidget.AnyWidget):
     # be overwritten by the short-hand options
     other_options = Dict(dict()).tag(sync=True)
 
-    view_reset = Bool(False).tag(sync=True) # Used for triggering a view reset
     view_download = Unicode(None, allow_none=True).tag(sync=True) # Used for triggering a download
     view_data = Array(default_value=None, allow_none=True, read_only=True).tag(sync=True, **ndarray_serialization)
     view_shape = List(None, allow_none=True, read_only=True).tag(sync=True)
@@ -223,6 +226,7 @@ class JupyterScatter(anywidget.AnyWidget):
 
     center_positions = List().tag(sync=True)
     dir_center_positions = List().tag(sync=True)
+    event_types = Dict(EVENT_TYPES, read_only=True).tag(sync=True)
 
     def __init__(self, data, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -231,14 +235,24 @@ class JupyterScatter(anywidget.AnyWidget):
         self.on_msg(self._handle_custom_msg)
 
     def _handle_custom_msg(self, event: dict, buffers):
-        if event["type"] == TOOLTIP_EVENT_TYPE and isinstance(self.data, pd.DataFrame):
+        if event["type"] == EVENT_TYPES["TOOLTIP"] and isinstance(self.data, pd.DataFrame):
             data = self.data.iloc[event["index"]]
             self.send({
-                "type": TOOLTIP_EVENT_TYPE,
+                "type": EVENT_TYPES["TOOLTIP"],
                 "index": event["index"],
                 "preview": data[event["preview"]] if event["preview"] is not None else None,
                 "properties": data[event["properties"]].to_dict()
             })
+
+    def show_tooltip(self, point_idx):
+        data = self.data.iloc[point_idx]
+        self.send({
+            "type": TOOLTIP_EVENT_TYPE,
+            "show": True,
+            "index": point_idx,
+            "preview": data[self.tooltip_preview] if self.tooltip_preview is not None else None,
+            "properties": data[self.tooltip_properties_non_visual_info.keys()].to_dict() if self.tooltip_properties_non_visual_info is not None else {}
+        })
 
     def create_download_view_button(self, icon_only=False, width=None):
         button = widgets.Button(
@@ -270,20 +284,33 @@ class JupyterScatter(anywidget.AnyWidget):
         button.on_click(click_handler)
         return button
 
-    def reset_view(self):
-        self.view_reset = True
+    def reset_view(self, animation: int = 0, data_extent: bool = False):
+        if data_extent:
+            self.send({
+                "type": EVENT_TYPES["VIEW_RESET"],
+                "area": {
+                    "x": self.points[:, 0].min(),
[…]
