+++ pyproject.toml
@@ -15,11 +15,11 @@ classifiers = [
     "Intended Audience :: Developers",
     "Intended Audience :: Science/Research",
     "Topic :: Multimedia :: Graphics",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 keywords = [
     "scatter",
@@ -32,7 +32,7 @@ keywords = [
     "widgets",
 ]
 dependencies = [
-    "anywidget>=0.2.0",
+    "anywidget>=0.9.0",
     "ipywidgets>=7.6,<9",
     "ipython",
     "matplotlib",
@@ -81,7 +81,7 @@ path = "js"
 
 [tool.hatch.envs.default]
 features = ["dev"]
-python = "3.11"
+python = "3.12"
 
 [tool.hatch.envs.default.scripts]
 test = "pytest ."

[…]
