### 2024-08-08 51b58ab
tell eslint to ignore console.logs
