### 2024-08-08 0887018
added directional search button and detect the corresponding clicks
