### 2024-08-09 flekschas/jupyter-scatter#146 @xiaohk
Review for the JOSS submission
Hello! I'm honored to review the JOSS submission for this cool package. It performs extremely well with large-scale datasets. I particularly love the linking scatter plot feature! Here are some comments:

### Usability feedback

1. The plot doesn't seem to show in the VS Code notebook. The paper mentions that jupyter-scatter supports VS Code, so are there specific installation instructions users should follow to make it work? Adding support for VS Code notebook, which is gaining popularity, could be valuable. If adding support is challenging, consider removing the claim from the introduction.

2. It would be helpful to have tooltips for the buttons in the menu bar on the left. While I can see the HTML title fields for the first three buttons in Jupyter Lab (not the last three though), I can't see any tooltips in Colab. Tooltips would benefit beginners using this tool.

3. After experimenting with it, I'm unsure of what the "activate rotation" button and camera button do.

4. I think the default background color is set to "white", but the downloaded image preview has a transparent background.

### Minor comments

1. There is a mysterious delay after selecting a new mode in Google Colab. For instance, I have to wait 3-5 seconds to switch between the pan & zoom mode and the lasso selection mode. I don't experience this delay in Jupyter Lab.

### Related work

1. In the paper, it can be valuable to discuss how jupyter-scatter differs from scalable embedding visualization tools such as [WizMap](https://github.com/poloclub/wizmap). While WizMap is designed for exploring and interpreting large-scale embeddings, it can also visualize scatter plots with millions of data points within computational notebooks.
