### 2024-08-09 #7059 @xiaohk
Thank you so much! Jupyter Scatter is such a nice tool to visualize large-scale scatter plots in computational notebooks! The software and paper both look pretty good. I have left a few comments at https://github.com/flekschas/jupyter-scatter/issues/146. I will complete my check list after seeing the authors' reply. Thank you!
