### 2024-08-13 #7059 @paulbrodersen
> These are not direct competitors because Jupyter Scatter is a general-purpose tool for rendering scatterplots in Notebook-like environments.

This is a bit like saying "I am a hammer, not a mallet, so I don't need to talk about mallets in my product description." However, if many people (including yourself!) use your hammer as a mallet, then I think you have to do some due diligence and discuss mallets, too. Notably, @xiaohk raises exactly the same point, and even points to the same competing software offering.
