### 2024-08-13 0bc4f7f
feat: improve view save button
The serializer now takes care of the nd array shape so you can directly use `scatter.widget.view_data` as the image data.
