### 2024-08-13 3837990
docs: describe how to export an image
