### 2024-08-15 cefbaf2
fix: hide rotation button
Also update regl-scatterplot to v1.10.3
