### 2024-08-16 55268db
feat: add background color to exported view
