### 2024-08-16 7624a90
fix: add a paragraph on how jscatter compares to wizmap and datamapplot
