+++ paper/refs.bib
@@ -261,3 +261,21 @@ year = {2022}
   year = {2015},
   url = {https://github.com/jupyter-widgets/ipywidgets}
 }
+
+@article{wang2023wizmap,
+  title = {{{WizMap}}: {{Scalable Interactive Visualization}} for {{Exploring Large Machine Learning Embeddings}}},
+  shorttitle = {{{WizMap}}},
+  author = {Wang, Zijie J. and Hohman, Fred and Chau, Duen Horng},
+  year = {2023},
+  url = {http://arxiv.org/abs/2306.09328},
+  urldate = {2023-06-16},
+  archiveprefix = {arxiv},
+  journal = {arXiv 2306.09328}
+}
+
+@misc{datamapplot,
+  author = {McInnes, Leland and others},
+  title = {{DataMapPlot}: Creating beautiful plots of data maps},
+  year = {2024},
+  url = {https://github.com/TutteInstitute/datamapplot}
+}
