### 2024-08-16 4d56002
docs: add Embedding Projector to related work
