+++ paper/refs.bib
@@ -270,7 +270,8 @@ year = {2022}
   url = {http://arxiv.org/abs/2306.09328},
   urldate = {2023-06-16},
   archiveprefix = {arxiv},
-  journal = {arXiv 2306.09328}
+  journal = {arXiv 2306.09328},
+  doi = {10.48550/arXiv.2306.09328}
 }
 
 @misc{datamapplot,
@@ -278,4 +279,15 @@ year = {2022}
   title = {{DataMapPlot}: Creating beautiful plots of data maps},
   year = {2024},
   url = {https://github.com/TutteInstitute/datamapplot}
+}
+
+@article{smilkov2016embedding,
+  title = {Embedding Projector: Interactive Visualization and Interpretation of Embeddings},
+  author = {Daniel Smilkov and Nikhil Thorat and Charles Nicholson and Emily Reif and Fernanda B. Viégas and Martin Wattenberg},
+  year = {2016},
+  eprint = {1611.05469},
+  archivePrefix = {arXiv},
+  primaryClass = {stat.ML},
+  url = {https://arxiv.org/abs/1611.05469},
+  doi = {10.48550/arXiv.1611.05469},
 }
