### 2024-08-16 #7059 @flekschas
Thanks a lot for the review @xiaohk. I've responded in your ticket to the issues you've found. Let me know if that addresses your concerns.

@paulbrodersen I've added a paragraph comparing JScatter to Jay's [WizMap](https://github.com/poloclub/wizmap) and Leland's [DataMapPlot](https://datamapplot.readthedocs.io/en/latest/), and Google's [Embedding Projector](https://projector.tensorflow.org/). See https://github.com/flekschas/jupyter-scatter/pull/145. Other tools mention in the WizMap paper are either plain web applications that do not integrate or run at all in a Notebook-like environment (DeepScatter) or are not even available as a software package (Latent Space Cartography, embComp, Parallel Embeddings, EmbeddingVis, "Visual Exploration of Semantic Relationships in Neural Word Embeddings"). Let me know if that addresses your concerns.

Besides this, did you have a chance to look at the [other changes I made](https://github.com/openjournals/joss-reviews/issues/7059#issuecomment-2271614507) in regards to your review @paulbrodersen? Do the changes address your concerns?

If this the changes look good, I'll go ahead, merge the PRs and release a new version with all the fixes.

Thanks again for finding and reporting the bugs and glitches!
