### 2024-08-18 flekschas/jupyter-scatter#146 @xiaohk
> Thank you so much for the review! I'll address/discuss each point individually below:
>
> ### Usability feedback
> 1. Thank you for bringing this up. Jupyter Scatter generally works with VSCode out of the box. However, a few versions ago VSCode introduced an issue when used with macOS: [Plots won't display when using jupyter notebook in vscode #37 (comment)](https://github.com/flekschas/jupyter-scatter/issues/37#issuecomment-2256593011). I finally found a fix and JScatter should work as expected in VSCode. See [Plots won't display when using jupyter notebook in vscode #37](https://github.com/flekschas/jupyter-scatter/issues/37)
> 2. Good catch! I've added tooltips for the three lower buttons. See PR [feat: improve view export and save #147](https://github.com/flekschas/jupyter-scatter/pull/147)
> 3. Another good catch! The camera button wasn't documented anywhere, so I've added a new section on how to export images to the docs. See PR [feat: improve view export and save #147](https://github.com/flekschas/jupyter-scatter/pull/147). Regarding the rotation button, I've decided to hide it for now. I found and fixed the bug that prevented the rotation from working but then I realized that when axes are displayed, the rotation is creating a mismatch with the axes as the axes remain unchanged. See PR [fix: hide rotation button #148](https://github.com/flekschas/jupyter-scatter/pull/148)
> 4. Another good find. That was because the background color is not actually rendered with WebGL but [accomplished with CSS](https://github.com/flekschas/jupyter-scatter/blob/main/js/src/index.js#L1740-L1747). The image export is literally only an image of the canvas element whose background is transparent. I added functionality to include the background color in the exported and saved views. See PR [feat: improve view export and save #147](https://github.com/flekschas/jupyter-scatter/pull/147)
>
> ### Minor comments
> 1. Could you open a ticket? I just tried to reproduce it and I didn't notice any delay. Maybe I'm missing something?
>
>  Screen.Recording.2024-08-16.at.9.40.50.AM.mp4
>
> ### Related work
> 1. That's a good point. I've looked through the related work in your WizMap paper and most tool you cite are web applications or just visualization methods. The only other comparable embedding plot tool that's specifically designed to work in Jupyter Notebooks I found is DataMapPlot. So I ended up adding a paragraph to the _State of Field_ section in the paper that's compares JScatter to WizMap and DataPlotMap. Does the paragraph accurately reflect WizMap and JScatter in your opinion? See [docs: improve summary and state of field #145](https://github.com/flekschas/jupyter-scatter/pull/145)

Thank you so much for the fixes that addressed all the mentioned issues and for the detailed replies! I have no further concerns. Very exciting tool and great paper!
