### 2024-08-19 cead377
fix: hide rotation button (#148)
* fix: hide rotation button

Also update regl-scatterplot to v1.10.3

* fix: properly activate rotate mouse mode
