### 2024-08-19 a71f7f6
feat: improve view export and save (#147)
* feat: improve view save button

The serializer now takes care of the nd array shape so you can directly use `scatter.widget.view_data` as the image data.

* docs: describe how to export an image

* docs: update changelog

* feat: add background color to exported view
