+++ js/src/index.js
@@ -7,7 +7,7 @@ import { format } from 'd3-format';
 import { scaleLinear } from 'd3-scale';
 import { select } from 'd3-selection';
 
-import { Numpy1D, Numpy2D } from "./codecs";
+import { Numpy1D, Numpy2D, NumpyImage } from "./codecs";
 import { createHistogram } from "./histogram";
 import { createLegend } from "./legend";
 import {
@@ -23,6 +23,8 @@ import {
   createElementWithClass,
   remToPx,
   createTimeFormat,
+  addBackgroundColor,
+  imageDataToCanvas,
 } from "./utils";
 
 import { version } from "../package.json";
@@ -95,7 +97,6 @@ const properties = {
   connectionOpacityBy: 'pointConnectionOpacityBy',
   connectionSize: 'pointConnectionSize',
   connectionSizeBy: 'pointConnectionSizeBy',
-  viewDownload: 'viewDownload',
   viewSync: 'viewSync',
   hovering: 'hovering',
   axes: 'axes',
@@ -388,7 +389,14 @@ class JupyterScatterView {
       if (event.index !== this.tooltipPointIdx && event.show !== true) return;
       this.tooltipDataHandlers(event)
       if (event.show) this.showTooltip(event.index);
+      return;
     }
+
+    if (event.type === this.eventTypes.VIEW_DOWNLOAD) {
+      this.viewDownload(event.transparentBackgroundColor)
+      return;
+    }
+
     if (event.type === this.eventTypes.VIEW_RESET) {
       if (!this.scatterplot) return;
       if (event.area) {
@@ -409,6 +417,12 @@ class JupyterScatterView {
           }
         );
       }
+      return;
+    }
+
+    if (event.type === this.eventTypes.VIEW_SAVE) {
+      this.viewSave(event.transparentBackgroundColor)
+      return;
     }
   }
 
@@ -1739,8 +1753,14 @@ class JupyterScatterView {
   // Helper
   colorCanvas() {
     if (Array.isArray(this.backgroundColor)) {
-      this.container.style.backgroundColor = 'rgb(' +
-        this.backgroundColor.slice(0, 3).map((x) => x * 255).join(',') + ')';
+      const rgbStr =
+        this.backgroundColor.slice(0, 3).map((x) => x * 255).join(',');
+
+      const colorStr = this.backgroundColor.length === 4
+        ? `rgba(${rgbStr}, ${this.backgroundColor[3]})`
+        : `rgb(${rgbStr})`;
+
+      this.container.style.backgroundColor = colorStr;
     } else {
       this.container.style.backgroundColor = this.backgroundColor;
     }
@@ -2338,25 +2358,23 @@ class JupyterScatterView {
     this.zoomToHandler(this.model.get('zoom_to'));
   }
 
-  viewDownloadHandler(target) {
-    if (!target) return;
-
-    if (target === 'property') {
-      const image = this.scatterplot.export();
-      this.model.set('view_data', image.data);
-      this.model.set('view_shape', [image.width, image.height]);
-      this.model.set('view_download', null);
-      this.model.save_changes();
-      return;
-    }
-
-    this.scatterplot.get('canvas').toBlob((blob) => {
+  viewDownload(transparentBackgroundColor) {
+    const image = this.scatterplot.export();
+    const finalImage = transparentBackgroundColor
+      ? image
+      : addBackgroundColor(image, this.backgroundColor);
+    imageDataToCanvas(finalImage).toBlob((blob) => {
       downloadBlob(blob, 'scatter.png');
-      setTimeout(() => {
-        this.model.set('view_download', null);
-        this.model.save_changes();
-      }, 0);
     });
+  }
+
[…]
