### 2024-08-19 #7059 @paulbrodersen
> I've added a paragraph comparing JScatter to Jay's [WizMap](https://github.com/poloclub/wizmap) and Leland's [DataMapPlot](https://datamapplot.readthedocs.io/en/latest/), and Google's [Embedding Projector](https://projector.tensorflow.org/). [...] Let me know if that addresses your concerns.

I think the paragraph strengthens the paper. It signals that even for very specific applications your tool does not need to hide from the more specialized competition.

> Besides this, did you have a chance to look at the https://github.com/openjournals/joss-reviews/issues/7059#issuecomment-2271614507 in regards to your review @paulbrodersen? Do the changes address your concerns?

They do. I was ticking off the remaining boxes as the PRs were coming in, but I probably should have been more explicit -- my apologies.

@faroit I also have no further concerns regarding the draft and join @xiaohk in their recommendation to accept the paper.

Congratulations to the authors!
