### 2024-08-19 d87c011
docs: improve summary and state of field (#145)
* Remove extensive use of `widget` in Summary and clarify "support for interactive point selections"

* Fix hanging sentence

* Update anywidget citation to point to the preprint

* fix: add a paragraph on how jscatter compares to wizmap and datamapplot

* docs: add Embedding Projector to related work
