+++ paper/refs.bib
@@ -208,11 +208,14 @@ year = {2022}
   url = {https://github.com/ipython/traitlets},
 }
 
-@misc{anywidget,
-  author = {Manz, Trevor},
-  title = {{Anywidget}: {Jupyter} Widgets Made Easy},
-  year = {2024},
-  url = {https://github.com/manzt/anywidget},
+@article{anywidget,
+ title = {anywidget: reusable widgets for interactive analysis and visualization in computational notebooks},
+ url = {osf.io/tw9sg},
+ doi = {10.31219/osf.io/tw9sg},
+ publisher = {OSF Preprints},
+ author = {Manz, Trevor and Abdennur, Nezar and Gehlenborg, Nils},
+ year = {2024},
+ month = {Jun}
 }
 
 @misc{geonames,
@@ -257,4 +260,34 @@ year = {2022}
   title = {ipywidgets: Interactive Widgets for the Jupyter Notebook},
   year = {2015},
   url = {https://github.com/jupyter-widgets/ipywidgets}
+}
+
+@article{wang2023wizmap,
+  title = {{{WizMap}}: {{Scalable Interactive Visualization}} for {{Exploring Large Machine Learning Embeddings}}},
+  shorttitle = {{{WizMap}}},
+  author = {Wang, Zijie J. and Hohman, Fred and Chau, Duen Horng},
+  year = {2023},
+  url = {http://arxiv.org/abs/2306.09328},
+  urldate = {2023-06-16},
+  archiveprefix = {arxiv},
+  journal = {arXiv 2306.09328},
+  doi = {10.48550/arXiv.2306.09328}
+}
+
+@misc{datamapplot,
+  author = {McInnes, Leland and others},
+  title = {{DataMapPlot}: Creating beautiful plots of data maps},
+  year = {2024},
+  url = {https://github.com/TutteInstitute/datamapplot}
+}
+
+@article{smilkov2016embedding,
+  title = {Embedding Projector: Interactive Visualization and Interpretation of Embeddings},
+  author = {Daniel Smilkov and Nikhil Thorat and Charles Nicholson and Emily Reif and Fernanda B. Viégas and Martin Wattenberg},
+  year = {2016},
+  eprint = {1611.05469},
+  archivePrefix = {arXiv},
+  primaryClass = {stat.ML},
+  url = {https://arxiv.org/abs/1611.05469},
+  doi = {10.48550/arXiv.1611.05469},
 }
