### 2024-08-19 #7059 @flekschas
@xiaohk @paulbrodersen Thanks again for your reviews! 🙏 It definitely helped us improve the paper and Jupyter Scatter.
