### 2024-08-28 #7059 @faroit
@flekschas i proofed the paper and created a little PR. When this is closed, we can continue with the post-review checklist [linked above](https://github.com release of jupyter-scatter to be drafted, create this before then.
