### 2024-08-28 d9523d8
feat: add support for line-based annotations (#151)
* feat: add support line-based annotations

* docs: describe `scatter.annotations()`
