+++ jscatter/__init__.py
@@ -9,5 +9,6 @@ except PackageNotFoundError:
     __version__ = "uninstalled"
 
 from .jscatter import Scatter, plot
+from .annotations import Line, HLine, VLine, Rect
 from .compose import compose, link
 from .color_maps import okabe_ito, glasbey_light, glasbey_dark
