+++ docs/.vitepress/config.ts
@@ -71,6 +71,7 @@ export default defineConfig({
           { text: 'Selections', link: '/selections' },
           { text: 'Link Multiple Scatter Plots', link: '/link-multiple-plots' },
           { text: 'Axes & Legends', link: '/axes-legends' },
+          { text: 'Annotations', link: '/annotations' },
           { text: 'Tooltip', link: '/tooltip' },
           { text: 'Scales', link: '/scales' },
           { text: 'Connected Scatterplots', link: '/connected-scatterplots' },
