### 2024-08-28 3d10f4d
feat: add support for line-based annotations (#151)
* feat: add support line-based annotations

* docs: describe `scatter.annotations()`
