### 2024-08-30 #7059 @flekschas
I published a new version ([v0.18.0](https://pypi.org/project/jupyter-scatter/0.18.0/)) with all [changes from the reviews](https://github.com/flekschas/jupyter-scatter/blob/main/CHANGELOG.md#v0180). The doi for the archived version is `10.5281/zenodo.13391017` ([link to Zenodo](https://zenodo.org/records/13391017)).
