### 2024-09-02 6280b11
fix: re-enable point transition and make it explicit
