+++ jscatter/jscatter.py
@@ -23,6 +23,7 @@ VALID_ENCODING_TYPES = [
     pd.api.types.is_string_dtype,
 ]
 DEFAULT_HISTOGRAM_BINS = 20
+DEFAULT_TRANSITION_POINTS_DURATION = 3000
 
 # An "undefined" value
 UNDEF = Undefined()
@@ -236,6 +237,9 @@ class Scatter():
         self._filtered_points_ids = None
         self._filtered_points_idxs = None
 
+        self._transition_points = True
+        self._transition_points_duration = DEFAULT_TRANSITION_POINTS_DURATION
+
         # We need to set the background first in order to choose sensible
         # default colors
         self._reticle_color = 'auto'
@@ -362,7 +366,7 @@ class Scatter():
         self._zoom_padding = 0.33
         self._zoom_on_selection = False
         self._zoom_on_filter = False
-        self._options = {}
+        self._regl_scatterplot_options = {}
 
         self.x(x, kwargs.get('x_scale', UNDEF))
         self.y(y, kwargs.get('y_scale', UNDEF))
@@ -481,9 +485,13 @@ class Scatter():
             kwargs.get('zoom_on_filter', UNDEF),
         )
         self.annotations(
-            kwargs.get('annotations', UNDEF)
+            kwargs.get('annotations', UNDEF),
         )
-        self.options(kwargs.get('options', UNDEF))
+        self.options(
+            kwargs.get('transition_points', UNDEF),
+            kwargs.get('transition_points_duration', UNDEF),
+            kwargs.get('options', UNDEF),
+        )
 
     def get_point_list(self):
         connect_by = bool(self._connect_by)
@@ -612,7 +620,6 @@ class Scatter():
 
             if 'skip_widget_update' not in kwargs:
                 self.update_widget('points', self.get_point_list())
-                self.update_widget('transition_points', animate == True)
                 self.update_widget('x_domain', self._x_domain)
                 self.update_widget('x_scale_domain', self._x_scale_domain)
                 self.update_widget('y_domain', self._y_domain)
@@ -646,6 +653,7 @@ class Scatter():
         self,
         x: Optional[Union[str, List[float], np.ndarray, Undefined]] = UNDEF,
         scale: Optional[Union[Scales, Tuple[float, float], LogNorm, PowerNorm, None, Undefined]] = UNDEF,
+        animate: Optional[Union[bool, Undefined]] = UNDEF,
         **kwargs
     ) -> Union[Scatter, dict]:
         """
@@ -658,6 +666,9 @@ class Scatter():
             or a string referencing a column in `data`.
         scale : {'linear', 'time', 'log', 'pow'}, tuple of floats, matplotlib.color.LogNorm, or matplotlib.color.PowerNorm, optional
             The x scale
+        animate : bool, optional
+            If `True` or if `self._transition_points` is `True`, the points will
+            transition smoothly.
         kwargs : optional
             Options which can be used to skip updating the widget when
             `skip_widget_update` is set to `True`
@@ -752,6 +763,8 @@ class Scatter():
                 self.update_widget('x_title', self._x_by)
                 self.update_widget('x_domain', self._x_domain)
                 self.update_widget('x_scale_domain', self._x_scale_domain)
+                _animate = self._transition_points if animate is UNDEF else animate
+                self.update_widget('transition_points', _animate != False)
                 self.update_widget('points', self.get_point_list())
 
         if any_not([x, scale], UNDEF):
@@ -772,6 +785,7 @@ class Scatter():
         self,
         y: Optional[Union[str, List[float], np.ndarray, Undefined]] = UNDEF,
[…]
