### 2024-09-02 073eb9b
test: fix pandas deprecation warning and remove unused `tooltip_histograms_bins` argument (#156)
