### 2024-09-02 5455267
test: fix pandas deprecation warning and remove unused `tooltip_histograms_bins` argument (#156)
