### 2024-09-04 8965d17
fix: do not set tooltip title if element is undefined
