### 2024-09-04 cc1dcb3
fix: revert to `pd.api.types.is_string_dtype`
It turns out that `pd.StringDtype.is_dtype` and `pd.api.types.is_string_dtype` are not equivalent...
