+++ tests/test_jscatter.py
@@ -327,10 +327,11 @@ def test_scatter_check_encoding_dtype(df: pd.DataFrame):
     check_encoding_dtype(pd.Series([1], dtype='int'))
     check_encoding_dtype(pd.Series([0.5], dtype='float'))
     check_encoding_dtype(pd.Series(['a'], dtype='string'))
+    check_encoding_dtype(pd.Series(['a']))
     check_encoding_dtype(pd.Series(['a'], dtype='category'))
 
     scatter = Scatter(data=df, x='a', y='b', color_by='group')
     check_encoding_dtype(scatter.color_data)
 
-    with pytest.raises(ValueError) as e_info:
+    with pytest.raises(ValueError):
         check_encoding_dtype(pd.Series(np.array([1+0j])))
