### 2024-09-05 f4fd024
fix: properly update axes grid upon resize
Also delay the resize by an animation frame
