### 2024-09-05 b525282
fix: reapply point color when `colorBy` gets set
