### 2024-09-05 13abbed
fix: regressions and smaller visual glitches (#157)
* fix: do not set tooltip title if element is undefined

* fix: revert to `pd.api.types.is_string_dtype`

It turns out that `pd.StringDtype.is_dtype` and `pd.api.types.is_string_dtype` are not equivalent...

* fix: properly update axes grid upon resize

Also delay the resize by an animation frame

* fix: reapply point color when `colorBy` gets set

* docs: update changelog
