+++ tests/test_jscatter.py
@@ -5,7 +5,7 @@ import pytest
 from functools import partial
 from matplotlib.colors import AsinhNorm, LogNorm, Normalize, PowerNorm, SymLogNorm
 
-from jscatter.jscatter import Scatter, component_idx_to_name
+from jscatter.jscatter import Scatter, component_idx_to_name, check_encoding_dtype
 from jscatter.utils import create_default_norm, to_ndc, TimeNormalize
 
 
@@ -322,3 +322,16 @@ def test_scatter_axes_labels(df: pd.DataFrame):
 
     scatter.axes(labels=['axis 1', 'axis 2'])
     assert scatter.widget.axes_labels == ['axis 1', 'axis 2']
+
+def test_scatter_check_encoding_dtype(df: pd.DataFrame):
+    check_encoding_dtype(pd.Series([1], dtype='int'))
+    check_encoding_dtype(pd.Series([0.5], dtype='float'))
+    check_encoding_dtype(pd.Series(['a'], dtype='string'))
+    check_encoding_dtype(pd.Series(['a']))
+    check_encoding_dtype(pd.Series(['a'], dtype='category'))
+
+    scatter = Scatter(data=df, x='a', y='b', color_by='group')
+    check_encoding_dtype(scatter.color_data)
+
+    with pytest.raises(ValueError):
+        check_encoding_dtype(pd.Series(np.array([1+0j])))
