### 2024-09-05 bd69ccc
fix: re-enable point transition and make it explicit (#155)
* fix: re-enable point transition and make it explicit

* docs: update changelog
