### 2024-09-05 3f4110c
fix: re-enable point transition and make it explicit (#155)
* fix: re-enable point transition and make it explicit

* docs: update changelog
