### 2024-09-06 #7059 @crvernon
:wave: @flekschas - I just need you to address the following in your paper before I move to accept for publication.  No need to generate another software release.  Just simply modify the paper.

- [ ] In the figure 1 caption, you state:  "with FAUST Greene et al. (2022)" which should be formatted as "with FAUST (Greene et al., 2022)"
- [ ] Ensure that items needing capitalization in your references are formatted correctly.  For example, in LINE 226, "python" should have a capital "P".  This can be maintained by placing curly brackets around any characters in your bib file that you want to maintain the formatting of.

Let me know when these changes have been made and I will move forward.
