### 2024-09-08 7e06995
docs: fix capitalization
