+++ paper/refs.bib
@@ -188,7 +188,7 @@ year = {2022}
 }
 
 @misc{bokeh,
-  title = {Bokeh: Python library for interactive visualization},
+  title = {{Bokeh}: {Python} library for interactive visualization},
   author = {{Bokeh development team}},
   year = {2018},
   url = {https://bokeh.pydata.org/en/latest/},
@@ -234,7 +234,7 @@ year = {2022}
   number = {32},
   pages = {1057},
   author = {Jacob VanderPlas and Brian Granger and Jeffrey Heer and Dominik Moritz and Kanit Wongsuphasawat and Arvind Satyanarayan and Eitan Lees and Ilia Timofeev and Ben Welsh and Scott Sievert},
-  title = {Altair: Interactive Statistical Visualizations for Python},
+  title = {{Altair}: Interactive Statistical Visualizations for {Python}},
   journal = {Journal of Open Source Software}
 }
 
