### 2024-09-08 62615ba
docs: fix footnote and citation typo
