### 2024-09-08 eb7ce91
remove leftover footnote
