### 2024-09-08 #7059 @flekschas
@crvernon I addressed the three issue.

Regarding the footnote syntax, it might be useful to update https://joss.readthedocs.io/en/latest/paper.html#footnotes as it seems that documentation is outdated.
